requirements.txt
# Core dependencies
torch>=2.0.0
transformers>=4.45.0
datasets>=2.14.0
accelerate>=0.20.0

# JOLA framework
jola

# Data processing
pandas>=1.5.0
numpy>=1.24.0

# Configuration
PyYAML>=6.0

# Logging and monitoring
wandb>=0.15.0
tensorboard>=2.13.0

# Performance optimizations
flash-attn>=2.3.0  # For A100 GPU optimization
xformers>=0.0.20   # Alternative attention optimization

# Utilities
tqdm>=4.65.0
psutil>=5.9.0

# Optional: For better performance
ninja>=1.11.0
